+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   date range.
 - The pre transfer grants export now include the projects for a supplied date
   range.
+- The loading time of the by local authority view has been improved, it has to
+  load all projects so is still a slow view.
 
 ## [Release-73][release-73]